breakdown slice file to smaller files
