+++ src/features/subgraph/subgraph.slice.ts
@@ -1,44 +1,14 @@
-import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
-import { captureException } from '@sentry/nextjs';
-import {
-	IBalances,
-	IInfinitePositionReward,
-	ITokenDistroInfo,
-	IUnipool,
-	IUniswapV2Pair,
-	IUniswapV3Pool,
-	IUniswapV3Position,
-	ZeroBalances,
-} from '@/types/subgraph';
-import { StakingType, StreamType, RegenFarmType } from '@/types/config';
-import { transformSubgraphData } from '@/lib/subgraph/subgraphDataTransform';
-import { SubgraphQueryBuilder } from '@/lib/subgraph/subgraphQueryBuilder';
-import { fetchSubgraph } from '@/services/subgraph.service';
+import { createSlice } from '@reduxjs/toolkit';
+import { ZeroBalances } from '@/types/subgraph';
 import config from '@/configuration';
+import {
+	fetchCurrentInfoAsync,
+	fetchXDaiInfoAsync,
+	fetchMainnetInfoAsync,
+} from './subgraph.thunks';
+import { ISubgraphState } from './subgraph.types';
 
-export interface ISubgraphState {
-	balances: IBalances;
-	tokenDistroInfo?: ITokenDistroInfo;
-	uniswapV3Pool?: IUniswapV3Pool;
-	[StakingType.GIV_LM]?: IUnipool;
-	[StakingType.BALANCER]?: IUnipool;
-	[StakingType.SUSHISWAP]?: IUnipool;
-	[StakingType.HONEYSWAP]?: IUnipool;
-	[StakingType.UNISWAPV2]?: IUnipool;
-	[StakingType.UNISWAPV3]?: IUnipool;
-	userNotStakedPositions: IUniswapV3Position[];
-	userStakedPositions: IUniswapV3Position[];
-	allPositions: IUniswapV3Position[];
-	uniswapV2EthGivPair?: IUniswapV2Pair;
-	infinitePositionReward?: IInfinitePositionReward;
-	infinitePosition?: IUniswapV3Position;
-	[StreamType.FOX]?: ITokenDistroInfo;
-	[StreamType.CULT]?: ITokenDistroInfo;
-	[RegenFarmType.FOX_HNY]?: IUnipool;
-	[RegenFarmType.CULT_ETH]?: IUnipool;
-}
-
-const defaultSubgraphValue: ISubgraphState = {
+export const defaultSubgraphValues: ISubgraphState = {
 	balances: ZeroBalances,
 	userNotStakedPositions: [],
 	userStakedPositions: [],
@@ -46,85 +16,11 @@ const defaultSubgraphValue: ISubgraphState = {
 };
 
 const initialState = {
-	currentValues: defaultSubgraphValue,
-	mainnetValues: defaultSubgraphValue,
-	xDaiValues: defaultSubgraphValue,
+	currentValues: defaultSubgraphValues,
+	mainnetValues: defaultSubgraphValues,
+	xDaiValues: defaultSubgraphValues,
 	status: 'idle',
 };
-
-const fetchMainnetInfo = async (userAddress = '') => {
-	try {
-		const response = await fetchSubgraph(
-			SubgraphQueryBuilder.getMainnetQuery(userAddress),
-			config.MAINNET_NETWORK_NUMBER,
-		);
-		return transformSubgraphData(response);
-	} catch (e) {
-		console.error('Error on query mainnet subgraph:', e);
-		captureException(e, {
-			tags: {
-				section: 'fetchMainnetSubgraph',
-			},
-		});
-		return defaultSubgraphValue;
-	}
-};
-
-const fetchXDaiInfo = async (userAddress = '') => {
-	try {
-		const response = await fetchSubgraph(
-			SubgraphQueryBuilder.getXDaiQuery(userAddress),
-			config.XDAI_NETWORK_NUMBER,
-		);
-		return transformSubgraphData(response);
-	} catch (e) {
-		console.error('Error on query xDai subgraph:', e);
-		captureException(e, {
-			tags: {
-				section: 'fetchxDaiSubgraph',
-			},
-		});
-		return defaultSubgraphValue;
-	}
-};
-
-export const fetchXDaiInfoAsync = createAsyncThunk(
-	'subgraph/fetchXDaiInfo',
-	async (userAddress: string) => {
-		const response = await fetchXDaiInfo(userAddress);
-		// The value we return becomes the `fulfilled` action payload
-		console.log(response);
-		return response;
-	},
-);
-
-export const fetchMainnetInfoAsync = createAsyncThunk(
-	'subgraph/fetchMainnetInfo',
-	async (userAddress: string) => {
-		const response = await fetchXDaiInfo(userAddress);
-		// The value we return becomes the `fulfilled` action payload
-		console.log(response);
-		return response;
-	},
-);
-
-interface ICurrentInfo {
-	userAddress: string;
-	chainId: number;
-}
-
-export const fetchCurrentInfoAsync = createAsyncThunk(
-	'subgraph/fetchCurrentInfo',
-	async (props: ICurrentInfo) => {
-		const response =
-			props.chainId === config.MAINNET_NETWORK_NUMBER
-				? await fetchMainnetInfo(props.userAddress)
-				: await fetchXDaiInfo(props.userAddress);
-		// The value we return becomes the `fulfilled` action payload
-		console.log(response);
-		return { response, chainId: props.chainId };
-	},
-);
 
 export const subgraphSlice = createSlice({
 	name: 'subgraph',
