+++ src/features/subgraph/subgraph.thunks.ts
@@ -1,0 +1,37 @@
+import { createAsyncThunk } from '@reduxjs/toolkit';
+import config from '@/configuration';
+import { fetchXDaiInfo, fetchMainnetInfo } from './subgraph.services';
+import { ICurrentInfo } from './subgraph.types';
+
+export const fetchXDaiInfoAsync = createAsyncThunk(
+	'subgraph/fetchXDaiInfo',
+	async (userAddress: string) => {
+		const response = await fetchXDaiInfo(userAddress);
+		// The value we return becomes the `fulfilled` action payload
+		console.log(response);
+		return response;
+	},
+);
+
+export const fetchMainnetInfoAsync = createAsyncThunk(
+	'subgraph/fetchMainnetInfo',
+	async (userAddress: string) => {
+		const response = await fetchXDaiInfo(userAddress);
+		// The value we return becomes the `fulfilled` action payload
+		console.log(response);
+		return response;
+	},
+);
+
+export const fetchCurrentInfoAsync = createAsyncThunk(
+	'subgraph/fetchCurrentInfo',
+	async (props: ICurrentInfo) => {
+		const response =
+			props.chainId === config.MAINNET_NETWORK_NUMBER
+				? await fetchMainnetInfo(props.userAddress)
+				: await fetchXDaiInfo(props.userAddress);
+		// The value we return becomes the `fulfilled` action payload
+		console.log(response);
+		return { response, chainId: props.chainId };
+	},
+);
